+++ moe/bandit/__init__.py
@@ -14,6 +14,10 @@
   object for all bandit epsilon subtypes.
 * :mod:`moe.bandit.ucb1`: :class:`~moe.bandit.ucb.UCB1`
   object for allocating bandit arms and choosing the winning arm based on UCB1 policy.
+* :mod:`moe.bandit.ucb1_tuned`: :class:`~moe.bandit.ucb.UCB1Tuned`
+  object for allocating bandit arms and choosing the winning arm based on UCB1-tuned policy.
+* :mod:`moe.bandit.ucb`: a base :class:`~moe.bandit.epsilon.UCB`
+  object for all bandit UCB subtypes.
 * :mod:`moe.bandit.linkers`: linkers connecting ``bandit`` components.
 
 compute the Bandit Epsilon arm allocation and choosing the arm to pull next.